+++ crates/model/src/config.rs
@@ -3,19 +3,20 @@
 
 //! This module provides a model configuration.
 
-use logjuicer_report::Source;
-use regex::RegexSet;
+use logjuicer_report::{Content, Source};
+use regex::{Regex, RegexSet};
 use serde::{Deserialize, Serialize};
 use std::path::PathBuf;
 use thiserror::Error;
 
 mod default_excludes;
 
-pub struct Config {
-    includes: Option<RegexSet>,
-    excludes: RegexSet,
-    skip_duplicate: bool,
-    pub ignore_patterns: RegexSet,
+/// The loaded user config
+pub enum Config {
+    /// A single global target config
+    Static(TargetConfig),
+    /// A list of target config to be matched with the target content
+    Matchers(Vec<(MatcherConfig, TargetConfig)>),
 }
 
 #[derive(Error, Debug)]
@@ -54,7 +55,50 @@ impl Config {
         Config::from_config_file(&cf)
     }
 
+    /// Convert the raw ConfigFile into a loaded Config
     fn from_config_file(cf: &ConfigFile) -> Result<Self, Error> {
+        match cf {
+            ConfigFile::Empty => Ok(Config::default()),
+            ConfigFile::Static(tcf) => TargetConfig::from_config_file(tcf).map(Config::Static),
+            ConfigFile::Matchers(xs) if xs.is_empty() => {
+                Err(Error::UnknownFormat("Target list is empty".into()))
+            }
+            ConfigFile::Matchers(xs) => xs
+                .iter()
+                .map(|tmf| {
+                    Ok((
+                        MatcherConfig::from_config_file(tmf)?,
+                        TargetConfig::from_config_file(&tmf.config)?,
+                    ))
+                })
+                .collect::<Result<Vec<_>, _>>()
+                .map(Config::Matchers),
+        }
+    }
+
+    /// Get a target config for the given targetr content
+    pub fn get_target_config(&self, target: &Content) -> Option<&TargetConfig> {
+        match self {
+            // When the config is static, use it directly.
+            Config::Static(tc) => Some(tc),
+            // Otherwise, find the matcher for this target
+            Config::Matchers(matchers) => matchers
+                .iter()
+                .find(|mc| mc.0.matches(target))
+                .map(|mc| &mc.1),
+        }
+    }
+}
+
+pub struct TargetConfig {
+    includes: Option<RegexSet>,
+    excludes: RegexSet,
+    skip_duplicate: bool,
+    ignore_patterns: RegexSet,
+}
+
+impl TargetConfig {
+    fn from_config_file(cf: &TargetConfigFile) -> Result<Self, Error> {
         let includes = if cf.includes.is_empty() {
             None
         } else {
@@ -77,7 +121,7 @@ impl Config {
         } else {
             cf.skip_duplicate
         };
-        Ok(Config {
+        Ok(TargetConfig {
             includes,
             excludes,
             skip_duplicate,
@@ -95,11 +139,42 @@ impl Config {
         !self.excludes.is_match(fp)
     }
 
+    pub fn is_ignored_line(&self, line: &str) -> bool {
+        self.ignore_patterns.is_match(line)
+    }
+
     pub fn new_skip_lines(&self) -> Option<crate::unordered::KnownLines> {
         if self.skip_duplicate {
             Some(crate::unordered::KnownLines::new())
         } else {
             None
+        }
+    }
+}
+
+pub struct MatcherConfig {
+    job_re: Option<Regex>,
+}
+
+impl MatcherConfig {
+    fn from_config_file(cf: &TargetMatcherFile) -> Result<Self, Error> {
+        let job_re = cf.match_job.as_ref().map(|s| Regex::new(s)).transpose()?;
+        Ok(MatcherConfig { job_re })
+    }
+
+    fn match_job(&self, name: &str) -> bool {
+        self.job_re
+            .as_ref()
+            .map(|job| job.is_match(name))
+            .unwrap_or(true)
+    }
+
+    fn matches(&self, content: &Content) -> bool {
+        match content {
+            Content::Zuul(build) => self.match_job(&build.job_name),
+            Content::LocalZuulBuild(_, build) => self.match_job(&build.job_name),
+            Content::Prow(build) => self.match_job(&build.job_name),
+            _ => true,
         }
     }
 }
@@ -111,8 +186,29 @@ impl Default for Config {
 }
 
 #[derive(Serialize, Deserialize)]
+#[serde(untagged)]
+enum ConfigFile {
+    Static(TargetConfigFile),
+    Matchers(Vec<TargetMatcherFile>),
+    Empty,
+}
+
+impl Default for ConfigFile {
+    fn default() -> Self {
+        ConfigFile::Static(TargetConfigFile::default())
+    }
+}
+
+#[derive(Serialize, Deserialize)]
 #[serde(deny_unknown_fields)]
-struct ConfigFile {
+struct TargetMatcherFile {
+    match_job: Option<String>,
+    config: TargetConfigFile,
+}
+
+#[derive(Serialize, Deserialize)]
+#[serde(deny_unknown_fields)]
+struct TargetConfigFile {
     #[serde(default)]
     includes: Vec<String>,
     #[serde(default)]
@@ -129,9 +225,9 @@ fn default_default_excludes() -> bool {
     true
 }
 
-impl Default for ConfigFile {
+impl Default for TargetConfigFile {
     fn default() -> Self {
-        ConfigFile {
+        TargetConfigFile {
             includes: Vec::new(),
             excludes: Vec::new(),
             default_excludes: true,
@@ -144,6 +240,7 @@ impl Default for ConfigFile {
 #[test]
 fn test_config_default_exclude() {
     let config = Config::default();
+    let config = config.get_target_config(&Content::sample("test")).unwrap();
     for src in [
         "config.yaml",
         "/config/.git/HEAD",
@@ -159,8 +256,7 @@ fn test_config_default_exclude() {
 fn test_config_default() {
     let config = config_from_yaml("");
     for src in ["service/api.log", "job-output.txt"] {
-        let source = Source::from_pathbuf(src.into());
-        assert_eq!(config.is_source_valid(&source), true,)
+        assert_eq!(config_check(&config, src), true);
     }
 }
 
@@ -171,6 +267,7 @@ fn config_from_yaml(yaml: &str) -> Config {
 
 #[cfg(test)]
 fn config_check(config: &Config, path: &str) -> bool {
+    let config = config.get_target_config(&Content::sample("test")).unwrap();
     config.is_source_valid(&Source::from_pathbuf(path.into()))
 }